Refactor SeleniumGrid and GridServer; upgrade JUnit Foundation
